fix: put get_blob_sidecars under SignedBeaconBlock
